!1427 RL training-inference precision alignment
* revert rejection_sampler.py
* fix npu_moe_finalize_routing output dtype
* !1305 92B训推精度对齐：mome add residual with bf16
* moe gate use fp32
* fix_bug
* add dsa_wi_weight_scal
* sort each row of topk_ids, reorder topk_weights, npu_moe_finalize_rout…
* lm_head fp32 and mtp logits fp32

+++ src/omni_npu/layers/fused_moe/prepare_permute_unpermute_finalize.py
@@ -171,15 +171,15 @@ class All2AllPrepPmtAndUnpmtFinal(FusedMoEPreparePermuteAndUnpermuteFinalize):
         gathered_tokens = new_x.new_empty(*expanded_x.shape)
         dist.all_to_all_single(gathered_tokens, new_x, input_splits, output_splits, group=get_ep_group().device_group)
         return torch_npu.npu_moe_finalize_routing(
-            gathered_tokens,
+            gathered_tokens.to(topk_weights.dtype),
             skip1=None,
             skip2=None,
             bias=None,
-            scales=topk_weights.to(gathered_tokens.dtype),
+            scales=topk_weights,
             expanded_src_to_dst_row=expanded_row_idx,
             export_for_source_row=None,
             drop_pad_mode=2,
-        )
+        ).to(gathered_tokens.dtype)
 
 
 class AGRSPrepPmtAndUnpmtFinal(FusedMoEPreparePermuteAndUnpermuteFinalize):
